Revert "can I revert to this now?"
This reverts commit 57acf1b3be820dde494ab0522a724e6987591b20.

+++ decoder/src/elf2table/mod.rs
@@ -217,7 +217,7 @@ pub fn parse_impl(elf: &[u8], check_version: bool) -> Result<Option<Table>, anyh
                 }
                 symbol::SymbolTag::Defmt(tag) => {
                     map.insert(
-                        entry.address() as usize,
+                        (entry.address() as u16) as usize,
                         TableEntry::new(
                             StringEntry::new(tag, sym.data().to_string()),
                             name.to_string(),
